requirements.txt
requests
beautifulsoup4
fake-useragent
markdown
jinja2
schedule
google-generativeai